requirements.txt
pyirsdk>=1.3.0
rich>=13.0.0
matplotlib>=3.7.0
numpy>=1.24.0
PySide6>=6.6
